Create Fabricating spherical carbon
To construct spherical carbon, you can choose the radius of the sphere, the number of atoms in the cut spherical cap, and you can switch the carbon atoms and replace them with other atoms. However, when replacing the atoms, please be aware of changing the bond lengths between the atoms.